Fixed iOS Chat List bug
Fixed the bug where the chat list in iOS ignored the active true/false filter so old ai chat rooms were showing in the list. Fixed by changing the behavior of saving old chat rooms and messages in the database but setting to inactive to instead simply deleting the data.